Adding autodocs of Pseudopotential class
